Revert "chore: move everything to 3 files - sterowanie, metoda, problem"
